completed update for release to maintainance teams.

+++ src/abb.sh
@@ -4,7 +4,7 @@
 bootstrap_ansible_update=0
 ansible_tests=0
 force_install=0
-exit=0
+verbose=0
 
 # sys
 title='Ansible Automation'
@@ -15,9 +15,6 @@ script_name=$(basename $(readlink -nf $0) ".sh")
 workspace=~/workspace
 installed="/etc/${script_name}.installed"
 
-# github
-github_url='https://github.com'
-
 # sysdeps
 pact_dependencies='figlet gcc-g++ wget python python-crypto python-paramiko p7zip libyaml-devel libffi-devel'
 pip_url='https://bootstrap.pypa.io/get-pip.py'
@@ -26,7 +23,7 @@ markupsafe_url='https://github.com/pallets/markupsafe/archive/master.zip'
 
 # ansible
 ansible_dir=/usr/local/etc/ansible
-ansible_repo_url="${github_url}/ansible/ansible.git"
+ansible_repo_url="https://github.com/ansible/ansible.git"
 ansible_version='stable-2.4'
 
 # console colours
@@ -34,7 +31,6 @@ green=\\e[92m
 normal=\\e[0m
 red=\\e[91m
 cyan=\\e[96m
-
 
 show_help()
 {
@@ -47,9 +43,34 @@ Run the ansible installer, with following options:
     -h  prints this help message
     -t  runs tests when initializing
     -u  run updates when initializing
+    -v  prints verbose log messages
 
 EOF
 }
+
+while getopts fhtuv opt
+do
+    case "$opt" in
+      f)  force_install=1;;
+      h)  show_help
+          exit 1;;
+      t)  ansible_tests=1;;
+      u)  bootstrap_ansible_update=1;;
+      v)  verbose=1;;
+      \?)		# unknown flag
+      	  show_help
+	  	  exit 1;;
+    esac
+done
+shift `expr $OPTIND - 1`
+
+if [ ${verbose} -eq 1 ]; then
+    # -- verbose enables
+    exec 4>&2 3>&1
+else
+    # -- verbose disabled, quiet all output
+    exec 4>/dev/null 3>/dev/null
+fi
 
 run()
 {
@@ -57,9 +78,10 @@ run()
     git checkout ${ansible_version} &> /dev/null
     source ./hacking/env-setup &> /dev/null
     sleep 1
+    cd ${workspace}
     clear
     figlet "${title}"
-    echo "\nConfigure Windows remotes with one of the below PS cmds\n. { iwr -useb https://bit.ly/2v9xsrG } | iex;\n. { iwr -useb http://mrmwebp.enbridge.com/OLFShare/ConfigureRemoteForAnsible.ps1 } | iex;\n. { iwr -useb https://raw.githubusercontent.com/ansible/ansible/devel/examples/scripts/ConfigureRemotingForAnsible.ps1 } | iex;\n"
+    echo "\nConfigure Windows remotes with the below PS cmd\n. { iwr -useb https://raw.githubusercontent.com/ansible/ansible/devel/examples/scripts/ConfigureRemotingForAnsible.ps1 } | iex;\n"
 }
 
 tests()
@@ -78,12 +100,14 @@ tests()
 
 update()
 {
-    if [ ${bootstrap_ansible_update} -eq 1 ]
+    if [ ! ${bootstrap_ansible_update} ]
     then
-        cd ${ansible_directory}
+        echo -n "Updating Ansible checkout.."
+        cd ${ansible_dir}
         git checkout devel &> /dev/null
         git pull --rebase &> /dev/null
         git submodule update --init --recursive &> /dev/null
+        echo "OK"
     fi
 }
 
@@ -101,6 +125,7 @@ mk_test_project ()
 {
     if [ ! -d  ${workspace}/test ]
     then
+        echo -n "Creating Test project..."
         mkdir -p ${workspace}/test/{conf,inventory}
         touch ${workspace}/test/conf/{vault_password,vault_key}
         chmod -x ${workspace}/test/conf/{vault_password,vault_key}
@@ -124,30 +149,46 @@ control_path = /tmp/ansible-ssh-%%h-%%p-%%r
 [privilege_escalation]
 become_user = true
 EOF
+        echo "OK"
     fi
 }
 
 sys_install()
 {
+    echo -n "Installing Pact Dependencies.."
     pact install ${pact_dependencies} &> /dev/null
+    echo "OK"
 
-    wget -q ${pip_url}
-    python get-pip.py &> /dev/null
-    rm -r get-pip.py
+    pip=$(which pip)
+    if [ $? -eq 1 ]
+    then
+        echo -n "Retrieving Python PIP.."
+        wget -q ${pip_url}
+        python get-pip.py &> /dev/null
+        rm -r get-pip.py
+        echo ".ok"
+    fi
 
+    echo -n "Installing PIP dependencies.."
     for dep in ${pip_dependencies}
     do
         pip install ${dep} --quiet &> /dev/null
     done
+    echo "OK"
 
-    pip uninstall markupsafe -y --quiet
-
-    cd ${workspace}
-    git clone https://github.com/pallets/markupsafe.git --quiet
-    cd markupsafe
-    python setup.py --without-speedups install &> /dev/null
-    cd ${workspace}
-    rm -rf markupsafe*
+    if [ ! -e /etc/marksupsafe.updated ]
+    then
+        echo -n "Replacing default markupsafe with source (github).."
+        pip uninstall markupsafe -y --quiet
+        cd ${workspace}
+        git clone https://github.com/pallets/markupsafe.git --quiet
+        cd markupsafe
+        python setup.py --without-speedups install &> /dev/null
+        cd ${workspace}
+        rm -rf markupsafe*
+        touch /etc/marksupsafe.updated
+        echo "OK"
+    fi
 }
 
 installer()
@@ -155,29 +196,31 @@ installer()
     if [ ! -f ${installed} ] || [ ${force_install} -eq 1 ]
     then
         cd ~
-        clear
         echo "${cyan}${title} Installer${normal}"
-        echo -n "installing please wait.."
 
         sys_install
 
-        if [ -d ${ansible_dir} ]
+        if [ ! -d ${ansible_dir} ]
         then
-            cd ${ansible_directory}
-            git checkout devel &> /dev/null
-            git pull --rebase &> /dev/null
-            git submodule update --init --recursive &> /dev/null
-        else
+            # echo -n "Updating Ansible source (github).."
+            # cd ${ansible_dir}
+            # git checkout devel &> /dev/null
+            # git pull --rebase &> /dev/null
+            # git submodule update --init --recursive &> /dev/null
+            # echo "OK"
+            echo -n "Retrieving Ansible source (github).."
             git clone ${ansible_repo_url} --recursive ${ansible_dir} --quiet
             cp ${ansible_dir}/examples/ansible.cfg ~/.ansible.cfg
             sed -i 's|#\?transport.*$|transport = paramiko|;s|#host_key_checking = False|host_key_checking = False|' ~/.ansible.cfg
-            touch ${installed}
+            echo "OK"
         fi
 
         mk_test_project
 
-        if [ `grep -q '# Ansible in Babun' ~/.zshrc` -eq 1 ]
+        abb=$(grep -q '# Ansible in Babun' ~/.zshrc)
+        if [ ! ${abb} ]
         then
+            echo -n "Updating ~/.zshrc (Babun).."
             cat >> ~/.zshrc << EOF
 
 #
@@ -192,7 +235,8 @@ then
 fi
 EOF
         fi
-        exit
+        touch ${installed}
+        echo ".OK"
     fi
 }
 
@@ -213,41 +257,8 @@ plugin()
 
 clear
 echo "${green}starting up..${normal}"
-while :; do
-    case $1 in
-        -h|-\?|--help)
-            show_help
-            exit=1
-            ;;
-        -f|--force)
-            force_install=1
-            ;;
-        -t|--test)
-            ansible_tests=1
-            ;;
-        -u|--update)
-            bootstrap_ansible_update=1
-            ;;
-        --)
-            shift
-            break
-            ;;
-        -?*)
-            printf 'WARN: Unknown option (ignored): %s\n' "$1" >&2
-            show_help
-            exit=1
-            ;;
-        *)
-            break
-    esac
-    shift
-done
-
-if [ ${exit} -eq 0 ]
-then
-    installer
-    update
-    plugin ${plugin_dir} "main"
-    run
-    tests
-fi
+installer
+update
+plugin ${plugin_dir} "main"
+run
+tests
